LayoutTests/imported/w3c/web-platform-tests/editing/other/recursive-exec-command-calls.tentative-expected.txt

CONSOLE MESSAGE: Error: assert_false: `execCommand("insertText") in the "input" event listener should return `false` expected false got true
CONSOLE MESSAGE: Error: assert_false: `execCommand("insertText") in the "input" event listener should return `false` expected false got true
CONSOLE MESSAGE: Error: assert_false: `execCommand("insertText") in the "input" event listener should return `false` expected false got true
CONSOLE MESSAGE: Error: assert_false: `execCommand("insertText") in the "input" event listener should return `false` expected false got true
CONSOLE MESSAGE: Error: assert_false: `execCommand("insertText") in the "input" event listener should return `false` expected false got true
CONSOLE MESSAGE: Error: assert_false: `execCommand("insertText") in the "input" event listener should return `false` expected false got true
CONSOLE MESSAGE: Error: assert_false: `execCommand("insertText") in the "input" event listener should return `false` expected false got true
CONSOLE MESSAGE: Error: assert_false: `execCommand("insertText") in the "input" event listener should return `false` expected false got true
CONSOLE MESSAGE: Error: assert_false: `execCommand("insertText") in the "input" event listener should return `false` expected false got true
0, 1, 2, 3, 4, 5, 6, 7, 8, 9

Harness Error (FAIL), message = Error: assert_false: `execCommand("insertText") in the "input" event listener should return `false` expected false got true

FAIL Recursive `Document.execCommand()` shouldn't be supported assert_equals: `execCommand("insertText") in the "input" event listener should do nothing expected "0" but got "0, 1, 2, 3, 4, 5, 6, 7, 8, 9"
